TEMMP

+++ src/tiaCleaner.cmd
@@ -1,4 +1,4 @@
-@echo off
+rem @echo off
 REM Keszitette: Tim Andras @ 2011
 title RENDSZER KARBANTARTµSA - tiaCleaner
 
@@ -17,7 +17,7 @@ rem ======MAIN==================================================================
 echo           (b rmikor le ll¡thatod a jobb fels‹ sarokban l‚v‹ 'X'-szel.)
 echo.
 
-if "%sdirLOG%" == "" set sdirLOG="%dirPRG%\dl\MyLogs"
+if "%sdirLOG%" == "" set sdirLOG=%dirPRG%\dl\MyLogs
 rem ---Karbantartas---
 if not exist "%dirPRG%\ver.txt" if exist "%dirPRG%\libs.zip" del "%dirPRG%\libs.zip"
 if not exist "%dirPRG%\dl" if exist "%dirPRG%\ver-local.txt" del "%dirPRG%\ver-local.txt"
@@ -25,42 +25,31 @@ if exist "%dirPRG%\libs.zip" ren "%dirPRG%\libs.zip" "%dirPRG%\libs_autodl.zip"
 
 echo * Friss¡t‚sek keres‚se...
 rem ---Jelenlegi verzio megallapitasa---
-if exist "%dirPRG%\ver-local.txt" (
-  set /p ver_local= < "%dirPRG%\ver-local.txt"
-) else (
-  set ver_local=n/a
-)
+if exist "%dirPRG%\ver-local.txt" set /p ver_local= < "%dirPRG%\ver-local.txt"
+else set ver_local=n/a
 echo      * Lok lis verzi¢: %ver_local%
 
 rem ---Cel verzio megallapitasa---
-set null="%temp%\tiac_ping.tmp"
-if not exist "%dirPRG%\ver.txt" (
-  ping %server% -n 1 > "%null%" && "%dirPRG%\bin\wget" "%urlVer%" -O "%dirPRG%\ver.txt" 2> "%null%"
-  title RENDSZER KARBANTARTµSA - tiaCleaner
-  if exist "%null%" del "%null%"
-)
-if exist "%dirPRG%\ver.txt" (
-  set /p ver_remote= < "%dirPRG%\ver.txt"
-) else (
-  set ver_remote=n/a
-)
+set null=%temp%\tiac_ping.tmp
+if not exist "%dirPRG%\ver.txt" ping %server% -n 1 > "%null%" && "%dirPRG%\bin\wget" "%urlVer%" -O "%dirPRG%\ver.txt" 2> "%null%"
+if not exist "%dirPRG%\ver.txt" title RENDSZER KARBANTARTµSA - tiaCleaner
+if not exist "%dirPRG%\ver.txt" if exist "%null%" del "%null%"
+
+if exist "%dirPRG%\ver.txt" set /p ver_remote= < "%dirPRG%\ver.txt"
+else set ver_remote=n/a
 echo      * T voli verzi¢: %ver_remote%
 rem ---Ellenorzes---
-if "%ver_local%%ver_remote%" == "n/an/a" (
-  if exist "%dirPRG%\ver.txt" del "%dirPRG%\ver.txt"
-  goto nincs
-)
-if not "%ver_local%" == "%ver_remote%" (
-  echo * Friss¡t‚s...
-  start /WAIT "" %0 #$INSTALLER$#
-  echo * Scriptk‚szlet felfriss¡tve
-)
+if "%ver_local%%ver_remote%" == "n/an/a" if exist "%dirPRG%\ver.txt" del "%dirPRG%\ver.txt"
+if "%ver_local%%ver_remote%" == "n/an/a" goto nincs
+if not "%ver_local%" == "%ver_remote%" echo * Friss¡t‚s...
+if not "%ver_local%" == "%ver_remote%" start /WAIT "" %0 #$INSTALLER$#
+if not "%ver_local%" == "%ver_remote%" echo * Scriptk‚szlet felfriss¡tve
 if exist "%dirPRG%\ver.txt" del "%dirPRG%\ver.txt"
 rem ---Futtatas---
 echo.
 echo.
 if not exist "%dirPRG%\dl\maxi_karbantartas.cmd" goto nincs
-set dirPRG="%dirPRG%\dl"
+set dirPRG=%dirPRG%\dl
 "%dirPRG%\maxi_karbantartas.cmd" %pcName% %*
 exit %errorlevel%
 
@@ -79,14 +68,8 @@ echo ---------------------------------------------------------------------------
 echo.
 set t=Jelenlegi  llapot biztons gi ment‚se... [ 1 / 5 ]
 title %t%
-set dirBACKUP="%dirPRG%\dl_backup_%date%"
-if exist "%dirBACKUP%" (
-  echo.
-  echo Nem lehet elv‚gezni a jelenlegi verzi¢ biztons gi ment‚s‚t, mert m r l‚tezik a "%dirBACKUP%" mappa!
-  echo.
-  pause
-  exit 1
-)
+set dirBACKUP=%dirPRG%\dl_backup_%date%
+if exist "%dirBACKUP%" goto installer_backup_error
 mkdir "%dirBACKUP%"
 if exist "%dirPRG%\dl" move "%dirPRG%\dl" "%dirBACKUP%\" || goto installer_error
 if exist "%dirPRG%\ver-local.txt" move "%dirPRG%\ver-local.txt" "%dirBACKUP%\" || goto installer_error
@@ -102,10 +85,8 @@ title %t%
 set pOK=0
 "%dirPRG%\bin\wget" %urlDL% -O "%dirPRG%\libs_autodl.zip" && set pOK=1
 title %t%
-if "%pOK%" == "0" (
-  if exist "%dirPRG%\libs_autodl.zip" del "%dirPRG%\libs_autodl.zip"
-  goto installer_error
-)
+if "%pOK%" == "0" if exist "%dirPRG%\libs_autodl.zip" del "%dirPRG%\libs_autodl.zip"
+if "%pOK%" == "0" goto installer_error
 
 :installer_unpack
 set t=Friss¡t‚s  kicsomagol sa... [ 4 / 5 ]
@@ -121,6 +102,7 @@ move /y "%dirPRG%\ver.txt" "ver-local.txt" || goto installer_error
 rmdir /s /q "%dirPRG%\unpack"
 del "%dirPRG%\libs_autodl.zip"
 rmdir /s /q "%dirBACKUP%"
+pause
 exit 0
 
 :installer_error
@@ -129,14 +111,19 @@ echo.
 echo HIBA LPETT FEL A K™VETKEZŠ FOLYAMAT ELVGZSE SORµN:
 echo %t%
 echo.
-if exist "%dirBACKUP%" (
-  echo Legutols¢ mûk”d‹  llapot vissza ll¡t sa... "%dirBACKUP%"
-  if exist "%dirPRG%\dl" rmdir /s /q "%dirPRG%\dl" || goto installer_restore_error
-  if exist "%dirPRG%\ver-local.txt" del "%dirPRG%\ver-local.txt" || goto installer_restore_error
-  if exist "%dirBACKUP%\dl" copy "%dirBACKUP%\dl" "%dirPRG%\" || goto installer_restore_error
-  if exist "%dirBACKUP%\ver-local.txt" copy "%dirBACKUP%\ver-local.txt" "%dirPRG%\" || goto installer_restore_error
-  echo     K‚sz
-)
+if exist "%dirBACKUP%" echo Legutols¢ mûk”d‹  llapot vissza ll¡t sa... "%dirBACKUP%"
+if exist "%dirBACKUP%" if exist "%dirPRG%\dl" rmdir /s /q "%dirPRG%\dl" || goto installer_restore_error
+if exist "%dirBACKUP%" if exist "%dirPRG%\ver-local.txt" del "%dirPRG%\ver-local.txt" || goto installer_restore_error
+if exist "%dirBACKUP%" if exist "%dirBACKUP%\dl" copy "%dirBACKUP%\dl" "%dirPRG%\" || goto installer_restore_error
+if exist "%dirBACKUP%" if exist "%dirBACKUP%\ver-local.txt" copy "%dirBACKUP%\ver-local.txt" "%dirPRG%\" || goto installer_restore_error
+if exist "%dirBACKUP%" echo     K‚sz
+echo.
+pause
+exit 1
+
+:installer_backup_error
+echo.
+echo Nem lehet elv‚gezni a jelenlegi verzi¢ biztons gi ment‚s‚t, mert m r l‚tezik a "%dirBACKUP%" mappa!
 echo.
 pause
 exit 1
@@ -146,4 +133,3 @@ echo     HIBA
 echo.
 pause
 exit 1
-
